requirements.txt
numpy==2.1.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==10.4.0
scipy==1.14.1
timm==1.0.9
torch==2.4.1
torchvision==0.19.1
